@@ -7,11 +7,11 @@
     "license": "MIT",
     "description": "The officially supported client for Postmark (http://postmarkapp.com)",
     "require": {
-        "php": ">=5.5.0",
+        "php": ">=7.0.0",
         "guzzlehttp/guzzle": "^6.0|^7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.4.0"
+        "phpunit/phpunit": "^6.0.0"
     },
     "autoload": {
         "psr-0": {
